Criando standalone com JOptionPane

+++ src/main/java/br/com/petra/dropbox/linkgen/services/MultiLinkService.java
@@ -9,21 +9,22 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.MediaType;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.ClientResponse;
-import org.springframework.web.reactive.function.client.ExchangeFilterFunction;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 import org.springframework.web.util.UriComponentsBuilder;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import javax.swing.*;
+import java.awt.*;
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
@@ -36,30 +37,11 @@ public class MultiLinkService {
 
     public static final ObjectMapper OBJECT_MAPPER = new ObjectMapper().enable(SerializationFeature.INDENT_OUTPUT);
 
-    private ExchangeFilterFunction logRequest() {
-        return (clientRequest, next) -> {
-            System.out.printf("Request: %s %s%n", clientRequest.method(), clientRequest.url());
-            clientRequest.headers()
-                    .forEach((name, values) -> values.forEach(value -> System.out.printf("%s=%s%n", name, value)));
-            return next.exchange(clientRequest);
-        };
-    }
-
-    private ExchangeFilterFunction logResponse() {
-        return ExchangeFilterFunction.ofResponseProcessor(clientResponse -> {
-            System.out.printf("Response: %s", clientResponse.headers().asHttpHeaders().get("property-header"));
-            return Mono.just(clientResponse);
-        });
-    }
-
     private final WebClient client = WebClient.builder()
-            .filter(logRequest())
-            .filter(logResponse())
             .baseUrl(ApplicationConstants.BASE_URL)
             .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .build();
 
-    @Async
     public void criarLinkSharedAll(String path, String apiKey) throws IOException {
         URI pathDropbox = DropboxLinkGenUtils.normalizarPath(path);
         String userHome = System.getProperty("user.home");
@@ -80,9 +62,24 @@ public class MultiLinkService {
         File txtLinkList = new File(pasta.getParentFile().toURI().resolve(pasta.getName() + "-links.txt"));
         System.out.printf("PASTA TEXTLINK: %s%n", txtLinkList.getAbsolutePath());
 
+        File[] files = ArrayUtils.nullToEmpty(pasta.listFiles(), File[].class);
+        int quantidadeArquivos = files.length;
+
+        JFrame loadingWindow = new JFrame("Carregando...");
+        JProgressBar progressBar = new JProgressBar(0, quantidadeArquivos); // Valor mínimo e máximo
+
+        loadingWindow.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        loadingWindow.setLayout(new BorderLayout());
+
+        loadingWindow.add(progressBar, BorderLayout.CENTER);
+
+        loadingWindow.pack();
+        loadingWindow.setLocationRelativeTo(null); // Centraliza a janela
+        loadingWindow.setVisible(true);
+
         FileUtils.writeStringToFile(txtLinkList, StringUtils.EMPTY, Charset.defaultCharset());
 
-        List<ResponseCreatedShareLinkDTO> linksGerados = Flux.from(DropboxLinkGenUtils.iterateFiles(pasta, pathDropbox))
+        List<ResponseCreatedShareLinkDTO> linksGerados = Flux.from(DropboxLinkGenUtils.iterateFiles(files, pathDropbox))
                 .delayElements(Duration.ofSeconds(5))
                 .flatMapSequential(dto ->
                         client.post()
@@ -93,6 +90,7 @@ public class MultiLinkService {
                                 .onStatus(HttpStatus.CONFLICT::equals, ClientResponse::createException)
                                 .bodyToMono(ResponseCreatedShareLinkDTO.class)
                                 .doOnNext(linkCriado -> DropboxLinkGenUtils.escreverLinksEmTxt(linkCriado, txtLinkList))
+                                .doOnEach(each -> progressBar.setValue(progressBar.getValue() + 1))
                                 .doOnError(error -> {
                                     if (error instanceof WebClientResponseException) {
                                         WebClientResponseException ex = (WebClientResponseException) error;
@@ -116,13 +114,21 @@ public class MultiLinkService {
                 .collectList()
                 .block();
 
+        loadingWindow.dispose();
 
+        int optionChosen = JOptionPane.showConfirmDialog(
+                null,
+                String.format("Process complete, %s links generated. Do you want to generate a JSON report?", CollectionUtils.emptyIfNull(linksGerados).size()),
+                "Success",
+                JOptionPane.YES_NO_OPTION,
+                JOptionPane.INFORMATION_MESSAGE
+        );
 
-        System.out.printf("Process complete, %s links generated", CollectionUtils.emptyIfNull(linksGerados).size());
-        String linksGeradosJSON = OBJECT_MAPPER.writeValueAsString(linksGerados);
-
-        File jsonLinkReport = new File(pasta.getParentFile().toURI().resolve(pasta.getName() + "-links-report.json"));
-        FileUtils.writeStringToFile(jsonLinkReport, linksGeradosJSON, Charset.defaultCharset());
+        if (optionChosen == JOptionPane.YES_OPTION) {
+            String linksGeradosJSON = OBJECT_MAPPER.writeValueAsString(linksGerados);
+            File jsonLinkReport = new File(pasta.getParentFile().toURI().resolve(pasta.getName() + "-links-report.json"));
+            FileUtils.writeStringToFile(jsonLinkReport, linksGeradosJSON, Charset.defaultCharset());
+        }
     }
 
 }
